BHAI Population Estimates

Total population-level estimates of healthcare-associated infection burden from BHAI analysis. This dataset contains parsed point estimates and 95% confidence intervals for cases, deaths, DALYs, YLL, and YLD.

Usage

bhai_pop_est

Format

A tibble with population estimates by infection and country:

infection

Type of healthcare-associated infection (or "ALL" for total)

cases_point_estimate

Point estimate for number of cases

cases_lower_ci

Lower 95% confidence interval for cases

cases_upper_ci

Upper 95% confidence interval for cases

deaths_point_estimate

Point estimate for number of deaths

deaths_lower_ci

Lower 95% confidence interval for deaths

deaths_upper_ci

Upper 95% confidence interval for deaths

daly_point_estimate

Point estimate for Disability-Adjusted Life Years

daly_lower_ci

Lower 95% confidence interval for DALYs

daly_upper_ci

Upper 95% confidence interval for DALYs

yll_point_estimate

Point estimate for Years of Life Lost

yll_lower_ci

Lower 95% confidence interval for YLL

yll_upper_ci

Upper 95% confidence interval for YLL

yld_point_estimate

Point estimate for Years Lived with Disability

yld_lower_ci

Lower 95% confidence interval for YLD

yld_upper_ci

Upper 95% confidence interval for YLD

country

Country or region (Germany or European Union)

population

Total population of the country or region

Source

Derived from BHAI simulations on German PPS 2011 and EU PPS data. Zacher et al. (2019) doi:10.2807/1560-7917.ES.2019.24.46.1900135

Details

This dataset is created by:

Running BHAI simulations (nsim = 500) on German and EU PPS data

Formatting results with bhai.prettyTable()

Parsing formatted strings to extract numeric estimates and confidence intervals

All estimates are population-level annual burden estimates. The "ALL" row represents the sum across all five HAI types.

References

Zacher B, Haller S, Eckmanns T, Noll I, Weiss B, Widders G, Gastmeier P, Monnet DL. Application of a new methodology and R package reveals a high burden of healthcare-associated infections (HAI) in Germany compared to the average in the European Union/European Economic Area, 2011 to 2012. Euro Surveill. 2019;24(46):pii=1900135.

See also

Examples

data(bhai_pop_est)

# View structure
str(bhai_pop_est)
#> tibble [12 × 18] (S3: tbl_df/tbl/data.frame)
#>  $ infection            : chr [1:12] "HAP" "SSI" "BSI" "UTI" ...
#>  $ cases_point_estimate : int [1:12] 107168 92425 26801 213207 36355 478745 112046 728270 882930 560543 ...
#>  $ cases_lower_ci       : int [1:12] 83237 75573 15827 177854 26405 425214 100059 693708 841222 532510 ...
#>  $ cases_upper_ci       : int [1:12] 136113 117042 42781 255501 50530 534626 124565 763192 931562 592104 ...
#>  $ deaths_point_estimate: int [1:12] 4113 2300 3841 3754 2064 16439 16959 26548 15078 12900 ...
#>  $ deaths_lower_ci      : int [1:12] 1093 1881 2015 1490 81 10602 10843 7358 6394 12254 ...
#>  $ deaths_upper_ci      : int [1:12] 8389 2909 6990 7533 4485 23202 23099 52492 30509 13626 ...
#>  $ daly_point_estimate  : int [1:12] 69477 28553 57993 68335 22101 253586 391871 555763 283437 176774 ...
#>  $ daly_lower_ci        : int [1:12] 37436 23286 31038 28234 2088 179459 263262 276780 124921 167937 ...
#>  $ daly_upper_ci        : int [1:12] 121253 35988 100293 127400 47754 352290 516010 845287 539080 187180 ...
#>  $ yll_point_estimate   : int [1:12] 42521 28108 49430 45406 21094 192603 333732 325095 192059 174132 ...
#>  $ yll_lower_ci         : int [1:12] 10974 22906 24875 18286 806 127665 213384 90101 81446 165424 ...
#>  $ yll_upper_ci         : int [1:12] 86817 35444 87104 95251 46369 275136 454546 642778 388617 183936 ...
#>  $ yld_point_estimate   : int [1:12] 28327 447 8688 20819 980 59949 56503 225409 87745 2721 ...
#>  $ yld_lower_ci         : int [1:12] 17915 343 4371 8584 178 40202 35012 141879 36034 2233 ...
#>  $ yld_upper_ci         : int [1:12] 44502 580 16377 40610 2141 86219 84795 312832 174551 3134 ...
#>  $ country              : chr [1:12] "Germany" "Germany" "Germany" "Germany" ...
#>  $ population           : int [1:12] 80767463 80767463 80767463 80767463 80767463 80767463 505529924 505529924 505529924 505529924 ...

# Total burden (ALL row)
if (FALSE) { # \dontrun{
bhai_pop_est |>
  dplyr::filter(infection == "ALL", country == "Germany")
} # }