BHAI PPS Sample Distribution

Sample distribution from Point Prevalence Surveys (PPS) showing the number of healthcare-associated infection (HAI) patients and total surveyed patients for Germany and the European Union.

Usage

bhai_pps_sample_distribution

Format

A data frame with one row per infection type and country:

country

Country or region name

infection

Type of healthcare-associated infection

num_hai_patients

Number of patients with HAI in the survey

num_survey_patients

Total number of patients surveyed

Source

Derived from BHAI package using German PPS 2011 and EU PPS data. Zacher et al. (2019) doi:10.2807/1560-7917.ES.2019.24.46.1900135

Details

The five infection types tracked are:

HAP: Hospital-acquired pneumonia

SSI: Surgical site infection

BSI: Bloodstream infection

UTI: Urinary tract infection

CDI: Clostridioides difficile infection

References

Zacher B, Haller S, Eckmanns T, Noll I, Weiss B, Widders G, Gastmeier P, Monnet DL. Application of a new methodology and R package reveals a high burden of healthcare-associated infections (HAI) in Germany compared to the average in the European Union/European Economic Area, 2011 to 2012. Euro Surveill. 2019;24(46):pii=1900135.

See also

Examples

data(bhai_pps_sample_distribution)

# View infection types
unique(bhai_pps_sample_distribution$infection)
#> [1] "BSI" "HAP" "UTI" "SSI" "CDI"

# Calculate prevalence
if (FALSE) { # \dontrun{
bhai_pps_sample_distribution |>
  dplyr::mutate(prevalence = num_hai_patients / num_survey_patients)
} # }