BHAI Stratified Summary Results

Detailed stratified results from BHAI (Burden of Healthcare-Associated Infections) analysis, including cases, deaths, and DALYs by country, infection type, sex, and age group, with 95% confidence intervals.

Usage

bhai_strata_summary

Format

A data frame with stratified results:

country

Country or region name

infection

Type of healthcare-associated infection

sex

Sex category (M/F)

age_group

Age group category (factor)

ncases_lower_ci

Lower 95% CI for number of cases

ncases_estimate

Median estimate for number of cases

ncases_upper_ci

Upper 95% CI for number of cases

ndeath_lower_ci

Lower 95% CI for number of deaths

ndeath_estimate

Median estimate for number of deaths

ndeath_upper_ci

Upper 95% CI for number of deaths

daly_lower_ci

Lower 95% CI for DALYs

daly_estimate

Median estimate for DALYs

daly_upper_ci

Upper 95% CI for DALYs

Source

Derived from BHAI simulations (nsim = 500) on German and EU PPS data. Zacher et al. (2019) doi:10.2807/1560-7917.ES.2019.24.46.1900135

Details

Stratification enables identification of demographic groups at highest risk. Age groups span from 0-1 years to 85-120 years. Results derived from 500 BHAI simulations incorporating uncertainty in prevalence, mortality, and disability weights.

References

Zacher B, Haller S, Eckmanns T, Noll I, Weiss B, Widders G, Gastmeier P, Monnet DL. Application of a new methodology and R package reveals a high burden of healthcare-associated infections (HAI) in Germany compared to the average in the European Union/European Economic Area, 2011 to 2012. Euro Surveill. 2019;24(46):pii=1900135.

See also

Examples

data(bhai_strata_summary)

# View age groups
levels(bhai_strata_summary$age_group)
#>  [1] "0-0"    "1-4"    "5-9"    "10-14"  "15-19"  "20-24"  "25-29"  "30-34" 
#>  [9] "35-39"  "40-44"  "45-49"  "50-54"  "55-59"  "60-64"  "65-69"  "70-74" 
#> [17] "75-79"  "80-84"  "85-120"