Flatten PPS Data Structure

Converts nested PPS (Point Prevalence Survey) S4 objects from BHAI into a flat tibble format suitable for analysis and visualization.

Usage

flatten_pps(pps_obj)

Arguments

pps_obj

An S4 object of class PPS from the BHAI package

Value

A tibble with one row per infection type, containing:

infection

Type of healthcare-associated infection

num_hai_patients

Number of patients with HAI

num_survey_patients

Total number of surveyed patients

length_of_stay

Average length of hospital stay

hospital_discharges

Matrix of hospital discharge data (list-column)

population

Population size

country

Country code or name

loi_pps

Length of infection PPS data (list-column)

num_hai_patients_by_stratum

HAI patients by stratum (list-column)

mccabe_scores_distr

Distribution of McCabe scores (list-column)

mccabe_life_exp

Life expectancy by McCabe score (list-column)

num_hai_patients_by_stratum_prior

Prior for HAI by stratum (list-column)

mccabe_by_stratum_prior

Prior for McCabe by stratum (list-column)

num_survey_patients_by_stratum

Survey patients by stratum (list-column)

bhai_options

BHAI options used (list-column)

bhai_summary

Summary statistics (list-column)

Examples

if (FALSE) { # \dontrun{
  # Flatten German PPS data
  german_df <- flatten_pps(BHAI::german_pps_conv)
  head(german_df)

  # Flatten EU PPS data
  eu_df <- flatten_pps(BHAI::eu_pps)
  head(eu_df)
} # }